Move gRASPAs into the raspa module
