+++ mofa/simulation/raspa/files/graspa_sycl_template/force_field_mixing_rules.def
