+++ mofa/simulation/raspa/files/raspa2_template/simulation.input
