+++ run_parallel_workflow.py
@@ -30,7 +30,7 @@ from mofa.selection.dft import DFTSelector
 from mofa.selection.md import MDSelector
 from mofa.simulation.cp2k import CP2KRunner, compute_partial_charges
 from mofa.simulation.mace import MACERunner
-from mofa.simulation.graspa_sycl import GRASPASyclRunner
+from mofa.simulation.raspa.graspa_sycl import GRASPASyclRunner
 from mofa.steering import GeneratorConfig, TrainingConfig, MOFAThinker, SimulationConfig
 from mofa.hpc.colmena import DiffLinkerInference
 from mofa.hpc.config import configs as hpc_configs, HPCConfig
